+++ public/index.html
@@ -94,10 +94,10 @@
   <script src="js/panel-resize-handler.js"></script>
   
   <!-- Vue Components -->
-  <script type="module" src="js/components/left-component.js"></script>
-  <script type="module" src="js/components/result-config-bar.js"></script>
+  <script type="module" src="js/components/left-input/container.js"></script>
+  <script type="module" src="js/components/right-config-bar.js"></script>
   <!-- Result Display Component (separated) -->
-  <script type="module" src="js/components/result-display/wrapper.js"></script>
-  <script type="module" src="js/components/global-config.js"></script>
+  <script type="module" src="js/components/right-display/container.js"></script>
+  <script type="module" src="js/components/top-config-bar.js"></script>
 </body>
 </html>
