+++ public/js/components/right-config-bar.js
