+++ public/js/components/top-config-bar.js
